Features in NoteService(when you don't give title and body of note for updating it, it will be deleted) and fixes

+++ src/main/java/org/example/tempnotes/notes/NoteService.java
@@ -21,6 +21,13 @@ public class NoteService {
     }
 
     public void addNote(NoteBody noteBody) {
+        if (
+                (noteBody.getTitle() == null || noteBody.getTitle().isEmpty())
+                    &&
+                (noteBody.getBody() == null || noteBody.getBody().isEmpty())
+        ) {
+            throw new IllegalArgumentException("Title or body must be given");
+        }
         Note note = new Note(
                 noteBody.getTitle(),
                 noteBody.getBody(),
@@ -30,22 +37,29 @@ public class NoteService {
     }
 
     public void deleteNote(String id) {
+        if (id == null) {
+            throw new IllegalArgumentException("The id mustn't be null");
+        }
         noteRepository.deleteById(id);
     }
 
     public void updateNote(NoteBody noteBody) throws IllegalArgumentException {
         if (noteBody.getId() == null) {
-            throw new IllegalArgumentException("the id field mustn't be null");
+            throw new IllegalArgumentException("The id field mustn't be null");
         }
-        if (noteBody.getTitle() == null) {
-            throw new IllegalArgumentException("the title field mustn't be null");
+        if (
+                (noteBody.getTitle() == null || noteBody.getTitle().isEmpty())
+                    &&
+                (noteBody.getBody() == null || noteBody.getBody().isEmpty())
+        ) {
+            deleteNote(noteBody.getId());
+        } else {
+            Note prevNote = getNote(noteBody.getId());
+
+            prevNote.setTitle(noteBody.getTitle());
+            prevNote.setBody(noteBody.getBody());
+            prevNote.setDestroyAtTime(LocalDate.parse(noteBody.getDestroyAtTime()));
+            noteRepository.save(prevNote);
         }
-
-        Note prevNote = getNote(noteBody.getId());
-
-        prevNote.setTitle(noteBody.getTitle());
-        prevNote.setBody(noteBody.getBody());
-        prevNote.setDestroyAtTime(LocalDate.parse(noteBody.getDestroyAtTime()));
-        noteRepository.save(prevNote);
     }
 }
